lidar_odometry: change parameter types from int to size_t and int64_t for consistency

+++ include/sycl_points/pipeline/lidar_odometry_params.hpp
@@ -22,16 +22,16 @@ struct Parameters {
     bool scan_downsampling_random_enable = true;
     size_t scan_downsampling_random_num = 5000;
 
-    int32_t scan_covariance_neighbor_num = 10;
+    size_t scan_covariance_neighbor_num = 10;
     bool scan_preprocess_box_filter_enable = true;
     float scan_preprocess_box_filter_min = 2.0f;
     float scan_preprocess_box_filter_max = 50.0f;
 
     float submap_voxel_size = 1.0f;
-    int32_t submap_covariance_neighbor_num = 10;
-    int32_t submap_color_gradient_neighbor_num = 10;
+    size_t submap_covariance_neighbor_num = 10;
+    size_t submap_color_gradient_neighbor_num = 10;
     float submap_max_distance_range = 30.0f;
-    int32_t submap_point_random_sampling_num = 2000;
+    size_t submap_point_random_sampling_num = 2000;
     float keyframe_inlier_ratio_threshold = 0.7f;
     float keyframe_distance_threshold = 2.0f;
     float keyframe_angle_threshold_degrees = 20.0f;
@@ -44,14 +44,14 @@ struct Parameters {
     float occupancy_grid_map_log_odds_limits_max = 4.0f;
     float occupancy_grid_map_occupied_threshold = 0.5f;
     bool occupancy_grid_map_enable_pruning = true;
-    uint32_t occupancy_grid_map_stale_frame_threshold = 100U;
+    size_t occupancy_grid_map_stale_frame_threshold = 100U;
 
     float registration_motion_prediction_factor = 0.5f;
     size_t registration_min_num_points = 100;
     bool registration_velocity_update_enable = true;
     size_t registration_velocity_update_iter = 1;
     bool registration_random_sampling_enable = true;
-    int32_t registration_random_sampling_num = 1000;
+    size_t registration_random_sampling_num = 1000;
     algorithms::registration::RegistrationParams reg_params;
 
     std::string odom_frame_id = "odom";
